Ultroid  v0.0.8 | 17-6-21

+++ plugins/carbon.py
@@ -1,5 +1,5 @@
 # Ultroid - UserBot
-# Copyright (C) 2020 TeamUltroid
+# Copyright (C) 2021 TeamUltroid
 #
 # This file is a part of < https://github.com/TeamUltroid/Ultroid/ >
 # PLease read the GNU Affero General Public License in
@@ -10,13 +10,14 @@
 
 • `{i}carbon <text/reply to msg/reply to document>`
     Carbonise the text with default settings.
+
 • `{i}rcarbon <text/reply to msg/reply to document>`
     Carbonise the text, with random bg colours.
 """
 
-import os
 import random
 
+import requests
 from carbonnow import Carbon
 
 from . import *
@@ -190,16 +191,19 @@ async def crbn(event):
             code = temp.message
     else:
         code = event.text.split(" ", maxsplit=1)[1]
-    carbon = Carbon(code=code)
-    xx = await carbon.save("ultroid_carbon")
+    webs = requests.get("https://carbonara.vercel.app/api/cook")
+    if webs.status_code == 502:
+        return await eor(
+            event, "`Temporary Server Error has Occured !\nPlease Try Again Later`"
+        )
+    carbon = Carbon(base_url="https://carbonara.vercel.app/api/cook", code=code)
+    xx = await carbon.memorize("ultroid_carbon")
     await xxxx.delete()
     await ultroid_bot.send_file(
         event.chat_id,
         xx,
         caption=f"Carbonised by [{OWNER_NAME}](tg://user?id={OWNER_ID})",
-        force_document=True,
     )
-    os.remove(xx)
 
 
 @ultroid_cmd(
@@ -220,16 +224,18 @@ async def crbn(event):
     else:
         code = event.text.split(" ", maxsplit=1)[1]
     col = random.choice(all_col)
-    carbon = Carbon(code=code, background=col)
-    xx = await carbon.save("ultroid_carbon")
+    webs = requests.get("https://carbonara.vercel.app/api/cook")
+    if webs.status_code == 502:
+        return await eor(
+            event, "`Temporary Server Error has Occured !\nPlease Try Again Later`"
+        )
+    carbon = Carbon(
+        base_url="https://carbonara.vercel.app/api/cook", code=code, background=col
+    )
+    xx = await carbon.memorize("ultroid_carbon")
     await xxxx.delete()
     await ultroid_bot.send_file(
         event.chat_id,
         xx,
         caption=f"Carbonised by [{OWNER_NAME}](tg://user?id={OWNER_ID})",
-        force_document=True,
     )
-    os.remove(xx)
-
-
-HELP.update({f"{__name__.split('.')[1]}": f"{__doc__.format(i=HNDLR)}"})
